Refactor: StringBuilder, null-safety, and parser fixes
Replace repeated String concatenations with StringBuilder, add @NotNull annotations, use Objects.equals for safe string comparison, and make several collections/fields final. Improve numeric casting for damage bonus, simplify control flow (early breaks), and convert Hero.buyUnits to a switch expression. Harden CommandParser (regex fix, safer resource handling, lambda for file filter) and streamline Parser iteration. Minor CommandCompiler catch cleanup. Add IntelliJ project files and compiled classes. These changes are primarily stylistic, safety and small-bug fixes.

+++ src/Units/Griffin.java
@@ -15,7 +15,7 @@ public final class Griffin extends Unit {
         }
         backed = true;
         double damage = (Core.Core.rng.nextInt(getStat("maxAttack")) + getStat("minAttack")) * getSize();
-        double dBonus = (getOwner().getStat("attack") * 10) / 100;
+        double dBonus = (double) (getOwner().getStat("attack") * 10) / 100;
         damage = damage + dBonus;
         unit.takeDamage(null, (int) Math.round(damage));
     }
